third_party/blink/web_tests/platform/android/external/wpt/pointerevents/capturing_boundary_event_handler_at_ua_shadowdom_touch-expected.txt

This is a testharness.js-based test.
[FAIL] Capturing boundary event handler at DIV
  assert_equals: received events with phases expected "pointerover-2,pointerenter-2,pointerout-2,pointerleave-2" but got ""
Harness: the test ran to completion.
